Add get user by id and moved it to directory models
I changed our function to get users_Id and change it location to make it clever
